+++ photutils/segmentation/core.py
@@ -56,10 +56,12 @@ class SegmentationImage:
     info : dict
         A dictionary containing auxiliary information about the
         segmentation image. For example, segmentation images returned
-        by :func:`~photutils.segmentation.deblend_sources` store any
-        deblending warnings under a ``'warnings'`` key. The dictionary
-        is empty if there is no auxiliary information. It is reset to an
-        empty dictionary when the ``data`` attribute is reassigned.
+        by :func:`~photutils.segmentation.deblend_sources` store the
+        input labels affected by deblending warnings under
+        ``'nonposmin_labels'`` and ``'n_markers_labels'`` keys. The
+        dictionary is empty if there is no auxiliary information. It is
+        reset to an empty dictionary when the ``data`` attribute is
+        reassigned.
 
     Notes
     -----